Collectable and Scoring System
The Collectable and Scoring System contains two scripts and 1 User interface.

ScoringSystem.cs: Update the UI score whenever the first person player collides with a collectable object.

CollectToken.cs: When the first person player collides with a collectable object, 50 points are added to the total score and the object is destroyed.

Note: Make sure the collectable object is set to "is triggered".